requirements.txt
Django==3.2.25
djangorestframework==3.14.0
psycopg2-binary==2.9.9
